more accurate messaging for --checkssh usage
figured it was something like this, but i was definitely nervous for a moment that i'd forgotten the `--checkssh` flag when i saw `Deploying tmp/repos/sul-dlss/dor_indexing_app` flash by on my terminal when i glanced back at it.

i included the `ssh_check` message because as it is, all you get is `Pseudo-terminal will not be allocated because stdin is not a terminal.` for each host.  a bit more inline info seemed like it'd be nice.