Allow deploying even when cocina-models versions vary
This commit changes the behavior of `bin/sdr` when deploying by loosening the requirement that all cocina-models versions are identical. We now allow cocina-models patch-level version divergence, and prompt the user whether to continue or not. If there is divergence at the major- or minor-level, we still halt the deploy.