Add ability to deploy & check specified branches in addition to tags
The SDR VMs will soon have MFA turned on for SSH in all environments. We have a solution for weekly mass deploys in place already that uses `sdr-infra` as a jump host, but currently if a developer needs to deploy a single branch for a single repo via capistrano, that requires N * M SSH connections: N being the number of VMs in a given environment and M being the number of steps in the capistrano deployment. Each of these connections will require a separate MFA interaction---which is untenable.

 This commit lets us extend our jump host solution to one-off deploys of codebases by allowing `sdr-deploy` to deal with both tags *and* branches.