Remove launchy
Now that we are deploying in a terminal on sdr-infra.stanford.edu the
launching of a browser pointed at nagios doesn't seem to work. I get
prompted for a cookie, and then when I answer "Y" it doesn't seem to
work. This commit prints the URL for clicking on instead.